+++ source/renderer/app/containers/settings/categories/SupportSettingsPage.js
@@ -1,17 +1,25 @@
 // @flow
-import React, { Component } from 'react';
+import React, { Component, Fragment } from 'react';
 import { observer, inject } from 'mobx-react';
 import { defineMessages, intlShape } from 'react-intl';
 import SupportSettings from '../../../components/settings/categories/SupportSettings';
 import type { InjectedProps } from '../../../types/injectedPropsType';
 import { generateFileNameWithTimestamp } from '../../../../../common/utils/files';
 import { getSupportUrl } from '../../../utils/network';
+import successIcon from '../../../assets/images/success-small.inline.svg';
+import NotificationMessage from '../../../components/widgets/NotificationMessage';
+import { DOWNLOAD_LOGS_SUCCESS_DURATION } from '../../../config/timingConfig';
 
 const messages = defineMessages({
   supportRequestLinkUrl: {
     id: 'settings.support.reportProblem.linkUrl',
     defaultMessage: '!!!https://iohk.zendesk.com/hc/en-us/requests/new/',
     description: '"Support request" link URL in the "Report a problem" section on the support settings page.',
+  },
+  downloadLogsSuccess: {
+    id: 'settings.support.reportProblem.downloadLogsSuccessMessage',
+    defaultMessage: '!!!Logs successfully downloaded',
+    description: 'Success message for download logs.',
   },
 });
 
@@ -24,6 +32,18 @@ export default class SupportSettingsPage extends Component<InjectedProps> {
 
   static defaultProps = { actions: null, stores: null };
 
+  constructor(props: any, context: any) {
+    super(props);
+    this.context = context;
+    this.registerOnDownloadLogsNotification();
+  }
+
+  componentWillUnmount() {
+    const { profile } = this.props.actions;
+    profile.downloadLogs.remove(this.openNotification);
+    this.closeNotification();
+  }
+
   handleSupportRequestClick = async (event: SyntheticEvent<HTMLButtonElement>) => {
     event.persist();
     const { intl } = this.context;
@@ -33,24 +53,63 @@ export default class SupportSettingsPage extends Component<InjectedProps> {
     this.props.stores.app.openExternalLink(supportUrl);
   };
 
+  openNotification = () => {
+    const { notifications } = this.props.actions;
+    const { id, duration } = this.notification;
+    notifications.open.trigger({ id, duration });
+  };
+
+  registerOnDownloadLogsNotification = () => {
+    const { profile } = this.props.actions;
+    profile.downloadLogs.listen(this.openNotification);
+  };
+
   handleDownloadLogs = () => {
     // TODO: refactor this direct access to the dialog api
     const fileName = generateFileNameWithTimestamp();
+    const { profile } = this.props.actions;
     const destination = global.dialog.showSaveDialog({
       defaultPath: fileName,
     });
     if (destination) {
-      this.props.actions.profile.downloadLogs.trigger({ fileName, destination, fresh: true });
+      profile.downloadLogs.trigger({ fileName, destination, fresh: true });
     }
   };
 
+  get notification() {
+    const { intl } = this.context;
+    return {
+      id: 'settings-page-download-logs-success',
+      duration: DOWNLOAD_LOGS_SUCCESS_DURATION,
+      message: intl.formatMessage(messages.downloadLogsSuccess),
+    };
+  }
+
+  closeNotification = () => {
+    const { id } = this.notification;
+    this.props.actions.notifications.closeActiveNotification.trigger({ id });
+  }
+
   render() {
+    const { stores } = this.props;
+    const { id, message } = this.notification;
     return (
-      <SupportSettings
-        onExternalLinkClick={this.props.stores.app.openExternalLink}
-        onSupportRequestClick={this.handleSupportRequestClick}
-        onDownloadLogs={this.handleDownloadLogs}
-      />
+      <Fragment>
+        <SupportSettings
+          onExternalLinkClick={stores.app.openExternalLink}
+          onSupportRequestClick={this.handleSupportRequestClick}
+          onDownloadLogs={this.handleDownloadLogs}
+        />
+        <NotificationMessage
+          icon={successIcon}
+          show={stores.uiNotifications.isOpen(id)}
+          onClose={this.closeNotification}
+          clickToClose
+          hasCloseButton
+        >
+          {message}
+        </NotificationMessage>
+      </Fragment>
     );
   }